move test utils into folder
