--- src/include/test/arrays
@@ -1,1 +1,0 @@
-../../external/data/arrays
