+++ src/include/test/unit_adj_list.cc
@@ -31,7 +31,7 @@
 
 #include <catch2/catch_all.hpp>
 #include "detail/graph/adj_list.h"
-#include "graphs/tiny.h"
+#include "test/utils/tiny_graphs.h"
 
 TEST_CASE("adj_list: test test", "[adj_list]") {
   REQUIRE(true);
