+++ src/include/test/unit_backwards_compatibility.cc
@@ -37,12 +37,12 @@
 #include "api/flat_l2_index.h"
 #include "api/ivf_flat_index.h"
 #include "api/vamana_index.h"
-#include "array_defs.h"
 #include "detail/linalg/matrix.h"
 #include "index/flat_l2_index.h"
 #include "index/ivf_flat_index.h"
 #include "index/vamana_index.h"
 #include "mdspan/mdspan.hpp"
+#include "test/utils/array_defs.h"
 #include "utils/print_types.h"
 
 TEST_CASE("backwards_compatibility: test test", "[backwards_compatibility]") {
