+++ src/include/test/unit_flat_l2_index.cc
@@ -31,8 +31,8 @@
  */
 
 #include <catch2/catch_all.hpp>
-#include "array_defs.h"
 #include "index/flat_l2_index.h"
+#include "test/utils/array_defs.h"
 
 TEST_CASE("index: test test", "[index]") {
   REQUIRE(true);
