+++ src/include/test/unit_ivf_qv.cc
@@ -31,12 +31,12 @@
 
 #include <catch2/catch_all.hpp>
 #include <cmath>
-#include "array_defs.h"
 #include "detail/ivf/dist_qv.h"  // dist_qv_finite_ram
 #include "detail/ivf/qv.h"
 #include "detail/linalg/matrix.h"
 #include "detail/linalg/tdb_io.h"
-#include "query_common.h"
+#include "test/utils/array_defs.h"
+#include "test/utils/query_common.h"
 
 TEST_CASE("qv: test test", "[qv]") {
   REQUIRE(true);
