+++ src/include/test/unit_nn-graph.cc
@@ -31,7 +31,7 @@
 
 #include <catch2/catch_all.hpp>
 #include "detail/graph/nn-graph.h"
-#include "query_common.h"
+#include "test/utils/query_common.h"
 
 bool global_debug = false;
