+++ src/include/test/unit_tdb_matrix.cc
@@ -37,7 +37,7 @@
 #include "detail/linalg/tdb_io.h"
 #include "detail/linalg/tdb_matrix.h"
 #include "mdspan/mdspan.hpp"
-#include "test/test_utils.h"
+#include "test/utils/test_utils.h"
 
 using TestTypes = std::tuple<float, double, int, char, size_t, uint32_t>;
